add dataContext

+++ Mishnat/Mishnat/Controllers/OrderController.cs
@@ -11,20 +11,20 @@ namespace Mishnat.Controllers
     public class OrderController : ControllerBase
     {
         readonly OrderService _orderService;
-         
         // GET: api/<OrderController>
         [HttpGet]
         public ActionResult<List<Order>> GetOrders()
         {
-           var result = _orderService.GetOrders();
+            var result = _orderService.GetOrders();
             if (result == null) return NotFound();
             return result;
         }
 
         // GET api/<OrderController>/5
         [HttpGet("{id}")]
-        public ActionResult<Order> GetOrderById(string id)
+        public ActionResult<Order> GetOrderById(int id)
         {
+            if (id < 0) return BadRequest();
             var result = _orderService.GetOrderById(id);
             if(result==null) return NotFound();
             return result;
@@ -34,20 +34,24 @@ namespace Mishnat.Controllers
         [HttpPost]
         public ActionResult<bool> AddOrder([FromBody] Order order)
         {
+            if(order == null) return BadRequest();
             return _orderService.AddOrder(order);
         }
 
         // PUT api/<OrderController>/5
         [HttpPut("{id}")]
-        public ActionResult<bool> UpdateOrder(string id, [FromBody] Order order)
+        public ActionResult<bool> UpdateOrder(int id, [FromBody] Order order)
         {
+
+            if(id < 0) return BadRequest();
             return _orderService.UpdateOrderById(id, order);
         }
 
         // DELETE api/<OrderController>/5
         [HttpDelete("{id}")]
-        public ActionResult<bool> DeleteOrder(string id)
+        public ActionResult<bool> DeleteOrder(int id)
         {
+            if (id < 0) return BadRequest();
             return _orderService.DeleteOrder(id);
         }
     }
