+++ Mishnat/Mishnat/Controllers/SellingController.cs
@@ -22,8 +22,9 @@ namespace Mishnat.Controllers
 
         // GET api/<SellingController>/5
         [HttpGet("{id}")]
-        public ActionResult<Selling> GetSellingById(string sellingId)
+        public ActionResult<Selling> GetSellingById(int sellingId)
         {
+            if (sellingId < 0) return BadRequest();
             var result = _sellingService.GetSellingById(sellingId);
             if(result==null) return NotFound();
             return result;
@@ -38,15 +39,19 @@ namespace Mishnat.Controllers
 
         // PUT api/<SellingController>/5
         [HttpPut("{id}")]
-        public ActionResult<bool> UpdateSelling(string sellingId, [FromBody] Selling selling)
+        public ActionResult<bool> UpdateSelling(int sellingId, [FromBody] Selling selling)
         {
+            if (sellingId < 0) return BadRequest();
+
             return _sellingService.UpdateSellingById(sellingId,selling);
         }
 
         // DELETE api/<SellingController>/5
         [HttpDelete("{id}")]
-        public ActionResult<bool> DeleteSelling(string sellingId)
+        public ActionResult<bool> DeleteSelling(int sellingId)
         {
+            if (sellingId < 0)
+                return BadRequest();
             return _sellingService.DeleteSelling(sellingId);
         }
     }
